fixed imports backports from main

+++ enmapbox/testing.py
@@ -99,7 +99,7 @@ class EnMAPBoxTestCase(TestCase):
         p = pathlib.Path(DIR_REPO) / 'test-outputs' / foldername
         if isinstance(subdir, str):
             p = p / subdir
-        if p.exists() and p.is_dir():
+        if cleanup and p.exists() and p.is_dir():
             shutil.rmtree(p)
         os.makedirs(p, exist_ok=True)
         return p
